requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
beanie>=1.26.0
motor>=3.5.1
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
Pillow>=10.4.0
loguru>=0.7.2
httpx>=0.27.2
pytest>=8.3.2
pytest-asyncio>=0.24.0
asgi-lifespan>=2.1.0
anyio>=4.4.0
ruff>=0.6.9
black>=24.10.0
